# openapi.yaml
openapi: 3.1.0
info:
  title: QuizForge API
  version: 0.5.0-dev
  summary: Asynchronous quiz and assessment tournaments.
  description: |
    QuizForge runs asynchronous quiz tournaments. A tournament opens for a
    window, players enter on their own schedule, and results roll up to a
    leaderboard. There is no synchronous "everyone answers at once" mode, by
    design.

    ## Status of this document

    This describes the API **exactly as it behaves today**.

    The shapes it first described were not all ones worth keeping. Six were
    corrected before anything consumed this document — while doing so was one
    commit rather than a breaking change — and are recorded in
    `docs/api/known-inconsistencies.md` with what they were and why they moved.

    ## Conventions

    - Every identifier is a prefixed `TypeId`: `acc_`, `wsp_`, `trn_`, `att_`,
      `qst_`. Bare UUIDs never appear in a response.
    - Every list endpoint returns `{ "data": [...], "nextCursor": ... }` and
      takes `limit` and `cursor`. Pagination is **keyset**, never offset: a
      cursor names the last row you saw, so a row inserted while you page
      through cannot duplicate or hide another one.
    - Out-of-range parameters are **rejected**, never silently corrected.

    Everything the product does is reachable here: workspaces, members, API
    keys, question banks, questions, tournaments and play.

    ## Questions are immutable

...
    question.

...

...

...

...
    have selected.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - bearerApiKey: []
paths:
...
  /v1/workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      summary: The workspaces you belong to
      description: |
        Every workspace the authenticated account is a member of, with the role
        it holds in each.

        Account-scoped: no workspace needs to be in scope, because this is how a
        client discovers which ids it may select.

        Requires a session. An API key already names exactly one workspace, so
        there is no set to enumerate.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: The caller's workspaces.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspacePage'
        '400':
          $ref: '#/components/responses/InvalidListRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '500':
          $ref: '#/components/responses/Internal'

    post:
      tags: [Workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      description: |
        Creates a workspace and makes the caller its `OWNER`.

        Account-scoped, necessarily: a new account belongs to no workspace, so
        requiring one in scope would make this unreachable and the account
        permanently unable to use the API.

        The slug is derived from the name and made unique; it is not accepted
        from the request, so a caller cannot squat on one.
      responses:
        '201':
          description: The workspace was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '500':
          $ref: '#/components/responses/Internal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
  /v1/workspaces/current:
    get:
      tags: [Workspaces]
      operationId: getCurrentWorkspace
      summary: The workspace currently in scope
      description: |
        The workspace named by `X-QuizForge-Workspace`, or implied by the API
        key, together with the caller's role in it.

        Useful for confirming which workspace a credential is acting in before
        doing anything destructive with it.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
      responses:
        '200':
          description: The workspace in scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/WorkspaceRequired'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '500':
          $ref: '#/components/responses/Internal'

    patch:
      tags: [Workspaces]
      operationId: renameWorkspace
      summary: Rename the workspace in scope
      description: |
        Changes the display name. Requires `OWNER`.

        The slug is **not** changed. It was unique when the workspace was
        created and anything already referring to it would break — renaming a
        workspace is a cosmetic act, not a re-identification.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameWorkspaceRequest'
      responses:
        '200':
          description: The workspace was renamed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qf_session
      description: |
        Opaque session token issued by `POST /v1/auth/login`, stored only as a
        SHA-256 digest on the server. Mutating requests authenticated this way
        are CSRF-protected and additionally require the `X-XSRF-TOKEN` header.
    bearerApiKey:
      type: http
      scheme: bearer
      description: |
        A workspace API key, sent as `Authorization: Bearer qf_…`. Shown once
        when created and stored only as a SHA-256 digest thereafter.

        Authenticates a workspace rather than a person, so endpoints that
        attribute an action to an account — starting and submitting attempts —
        reject it.
  parameters:
    WorkspaceHeader:
      name: X-QuizForge-Workspace
      in: header
      required: false
      description: |
        The workspace to operate on, as a `wsp_…` identifier.

        Required for every endpoint outside `/v1/auth` when authenticating with
        a session cookie; omitting it fails with `400 INVALID_REQUEST`. With an
        API key the workspace comes from the key and this header is ignored.

        Documented as `required: false` because it is genuinely optional for
        key-authenticated callers, not because it is safe to leave out.

        Naming a workspace you are not a member of returns
        **`403 PERMISSION_DENIED`**: the credential is valid, the workspace is
        not yours. A malformed value is `400 INVALID_REQUEST`.
      schema:
        $ref: '#/components/schemas/WorkspaceId'
    Limit:
      name: limit
      in: query
      required: false
      description: |
        How many items to return, 1–100.

        A value outside that range is `400 INVALID_REQUEST`. It is deliberately
        not clamped: answering a request for 1000 with 100 items is
        indistinguishable from a collection that only holds 100, so a client
        would conclude it had everything when it had a prefix.
      schema:
        type: integer
        format: int32
        default: 25
        minimum: 1
        maximum: 100
    PageCursor:
      name: cursor
      in: query
      required: false
      description: |
        The `nextCursor` from the previous page. Omit for the first page.

        **Treat this as opaque.** It is not a page number, an offset, or an
        identifier, and its encoding is not part of the contract. A value this
        API did not issue is `400 INVALID_CURSOR` rather than being quietly
        treated as "start again".
      schema:
        $ref: '#/components/schemas/CursorParameter'
...
  responses:
    InvalidRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspaceRequired:
      description: |
        No workspace is in scope. Carries `code: INVALID_REQUEST`.

        Sent when a session-authenticated request omits
        `X-QuizForge-Workspace`. Tenant isolation depends on a workspace being
        selected, so such a request is refused before it reaches a handler
        rather than being served unfiltered.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspaceForbidden:
      description: |
        The credential is valid but the account is not a member of the workspace
        it named. Carries `code: PERMISSION_DENIED`.

        Distinct from `401` deliberately: re-authenticating cannot fix this, so
        telling the caller they are unauthenticated would send them to the wrong
        remedy.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationRequired:
      description: |
        No valid credential was presented. Carries
        `code: AUTHENTICATION_REQUIRED`.

        Also returned in two cases that are not obviously authentication
        failures:

        - A path that does not exist. Authentication is evaluated before
          routing, so an anonymous caller learns nothing about which paths are
          real.
        It is **not** returned for a workspace you are not a member of — that
        is `403`, because the credential itself is valid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InvalidListRequest:
      description: |
        The request could not be paged. One of:

        - no workspace is in scope — `code: INVALID_REQUEST`
        - `limit` is outside 1–100 — `code: INVALID_REQUEST`
        - `cursor` was not issued by this API — `code: INVALID_CURSOR`

        The three are distinguished by `code`, which is why they share a status
        without sharing a meaning.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Internal:
      description: |
        Something failed unexpectedly. The cause is logged server-side and
        never included in the response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      title: Problem
      description: |
        An RFC 9457 Problem Details document. Every error in the API has this
        shape.

        Clients should branch on `code`. It is the only field with a stability
        guarantee: a code that exists will never change meaning or HTTP status,
        and new codes are added rather than existing ones repurposed.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          maxLength: 200
          description: Stable URI identifying the error class.
          examples: ['https://quizforge.dev/errors/not-found']
        title:
          type: string
          maxLength: 100
          description: The error code, repeated as a human-readable title.
          examples: ['NOT_FOUND']
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code, repeated in the body.
          examples: [404]
        detail:
          type: string
          maxLength: 2000
          description: |
            Human-readable explanation. **Not stable** — reworded freely. Never
            parse it.
          examples: ['no such attempt']
        instance:
          type: string
          format: uri-reference
          maxLength: 2000
          description: The path that produced the error.
          examples: ['/v1/attempts/att_01a0136fb3937209a5719e8fb7db26bf']
        code:
          $ref: '#/components/schemas/ErrorCode'
    ErrorCode:
      type: string
      title: ErrorCode
      description: |
        The machine-readable error code. Exhaustive: every error the API can
        produce carries one of these.

        Each maps to exactly one HTTP status, and that mapping is part of the
        contract.
      enum:
        - INVALID_REQUEST
        - INVALID_CREDENTIALS
        - AUTHENTICATION_REQUIRED
        - PERMISSION_DENIED
        - NOT_FOUND
        - ALREADY_EXISTS
        - RATE_LIMITED
        - INTERNAL
      examples: ['NOT_FOUND']
      x-error-codes:
        INVALID_REQUEST:
          status: 400
          meaning: Malformed input, a failed validation rule, or an operation the current state does not allow.
        INVALID_CREDENTIALS:
          status: 401
          meaning: The email or password was wrong. Which one is never disclosed.
        AUTHENTICATION_REQUIRED:
          status: 401
          meaning: No credential was presented, or it has expired or been revoked.
        PERMISSION_DENIED:
          status: 403
          meaning: Authenticated, but the role is insufficient for this action.
        NOT_FOUND:
          status: 404
          meaning: No such object, or it belongs to another tenant. Deliberately ambiguous.
        ALREADY_EXISTS:
          status: 409
          meaning: A uniqueness constraint would be violated.
        RATE_LIMITED:
          status: 429
          meaning: Too many requests. Not yet emitted — rate limiting arrives with M4.
        INTERNAL:
          status: 500
          meaning: An unhandled failure. The cause is logged and never returned.
...
    WorkspaceId:
      type: string
      title: WorkspaceId
      description: A workspace identifier. UUIDv7 rendered as 32 lowercase hex characters with a `wsp_` prefix.
      pattern: '^wsp_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['wsp_01a0136fb3937209a5719e8fb7db26bf']
...
    Cursor:
      type: [string, 'null']
      title: Cursor
      description: |
        Pass this back as `cursor` to get the next page. `null` means there is
        no next page — stop, rather than asking for one you already know is
        empty.

        **Opaque.** Not a page number, not an offset, not an identifier. Its
        encoding is not part of the contract and may change; a cursor issued by
        an older version will be rejected rather than misread.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
    CursorParameter:
      type: string
      title: CursorParameter
      description: An opaque cursor previously returned as `nextCursor`.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
...
    Role:
      type: string
      title: Role
      description: |
        What a member may do in a workspace, from most to least privileged.

        - `OWNER` — everything, including managing API keys and renaming the
          workspace. A workspace always has at least one.
        - `ADMIN` — manage members, content and tournaments.
        - `EDITOR` — manage content and tournaments.
        - `VIEWER` — read only.
      enum: [OWNER, ADMIN, EDITOR, VIEWER]
      examples: ['OWNER']
...
    Workspace:
      type: object
      title: Workspace
      description: A tenant. Every question, tournament and attempt belongs to exactly one.
      required: [id, name, slug, role]
      properties:
        id:
          $ref: '#/components/schemas/WorkspaceId'
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          maxLength: 140
          description: |
            Derived from the name at creation and unique across the system. Not
            changed by a rename, because anything already referring to it would
            break.
        role:
          $ref: '#/components/schemas/Role'
      examples:
        - id: wsp_01a0136fb3937209a5719e8fb7db26bf
          name: Acme Training
          slug: acme-training
          role: OWNER
    WorkspacePage:
      type: object
      title: WorkspacePage
      required: [data]
      properties:
        data:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/Workspace'
        nextCursor:
          $ref: '#/components/schemas/Cursor'
...
    CreateWorkspaceRequest:
      type: object
      title: CreateWorkspaceRequest
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: The slug is derived from this and is not accepted from the request.
      examples:
        - name: Acme Training
    RenameWorkspaceRequest:
      type: object
      title: RenameWorkspaceRequest
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
      examples:
        - name: Acme Training EMEA
...